third_party/blink/web_tests/http/tests/eventsource/workers/eventsource-csp-redirect.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy"
      content="connect-src http://127.0.0.1:8000/js-test-resources/js-test.js http://127.0.0.1:8000/resources/redirect.php">
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
startWorker("../script-tests/eventsource-csp-redirect.php");
</script>
</body>
</html>
