third_party/blink/web_tests/http/tests/eventsource/workers/eventsource-csp-redirect-expected.txt

CONSOLE ERROR: Refused to connect to 'http://127.0.0.1:8080/resources/redirect.php' because it violates the following Content Security Policy directive: "connect-src http://127.0.0.1:8000/js-test-resources/js-test.js http://127.0.0.1:8000/resources/redirect.php".

[Worker] Test that basic EventSource cross-origin requests fail on blocked CSP redirect.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-csp-redirect.php
PASS [Worker] es = new EventSource("http://127.0.0.1:8000/resources/redirect.php?code=307&cors_allow_origin=*&url=http://127.0.0.1:8080/resources/redirect.php") did not throw exception.
PASS [Worker] es.readyState is EventSource.CLOSED
PASS successfullyParsed is true

TEST COMPLETE
